fix in NameConf::rectifyDir method
