+++ DataFormats/Detectors/Common/include/DetectorsCommonDataFormats/NameConf.h
@@ -38,7 +38,7 @@ class NameConf
   static std::string getFullPath(const std::string_view p);
 
   // Rectify directory allowing empty one and adding '/' if it is not empty
-  static void rectifyDirectory(std::string& dir);
+  static std::string rectifyDirectory(const std::string& _dir);
 
   // The Hits file name are generated by hardcoded schema, only prefix is mutable to allow the embedding
   static std::string getHitsFileName(o2::detectors::DetID d, const std::string_view prefix = STANDARDSIMPREFIX)
